contacts finished | groups(almost) | a lot of problems

+++ public/js/index.js
@@ -8,15 +8,16 @@ class App {
 		this.c_id = '';
 		this.submit = document.getElementById('contact-submit');
 		this.addButton = document.getElementById('addButton');
+		this.groupButton = document.getElementById('groupButton');
 		this.deleteButton = document.getElementById('deleteButton');
 		this.add = true;
 	}
 	getData() {
 		// console.log(api.base_url);
-		api.getContacts().then((data) => this.displayNotes(data)).catch((err) => console.log(err));
+		api.getContacts().then((data) => this.displayContacts(data)).catch((err) => console.log(err));
 	}
 
-	displayNotes(contacts) {
+	displayContacts(contacts) {
 		// console.log();
 
 		ui.display(contacts);
@@ -32,7 +33,12 @@ class App {
 	}
 
 	deleteContact(id) {
-		api.deleteContact(id).then((data) => this.getData()).catch((err) => console.log(err));
+		if (confirm('Are you sure?'))
+			api.deleteContact(id).then((data) => this.getData()).catch((err) => console.log(err));
+	}
+
+	getGroups() {
+		api.getGroups().then((data) => ui.displayGroups(data)).catch((err) => console.log(err));
 	}
 }
 
@@ -45,6 +51,10 @@ const onLoad = () => {
 app.addButton.addEventListener('click', () => {
 	(app.name.value = ''), (app.number.value = ''), (app.nickname.value = '');
 });
+
+// app.groupButton.addEventListener('click', () => {
+// 	app.getGroups();
+// });
 app.submit.addEventListener('click', (e) => {
 	if (app.name.value !== '' && app.number.value !== '') {
 		if (app.add) {
